ExtratosScreen/Formaters

+++ Teste Solutis-Santander/Model/User.swift
@@ -9,11 +9,11 @@ import UIKit
 
 class User: NSObject {
     var nome:String
-    var saldo:Int
+    var saldo:Double
     var cpf:String
     var token:String
     
-    init(nome:String, saldo:Int,cpf:String,token:String) {
+    init(nome:String, saldo:Double,cpf:String,token:String) {
         self.nome = nome
         self.saldo = saldo
         self.cpf = cpf
